backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
sqlalchemy>=2.0.29
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
celery>=5.3.6
redis>=5.0.3
langchain>=0.1.13
langchain-openai>=0.1.1
openai>=1.14.2
faiss-cpu>=1.8.0
beautifulsoup4>=4.12.3
requests>=2.31.0
python-dotenv>=1.0.1
feedparser>=6.0.11
python-dateutil>=2.9.0
moviepy>=1.0.3
gtts
pillow
replicate
langchain-community
aiosqlite
yfinance
sentence-transformers
